ProductFacade::saveParameters - move extended logic from project-base to the framework package
- ParameterRepository::findOrCreateParameterValueByValueTextAndLocale() replaced with findOrCreateParameterValueByParameterValueData()
